Improved AppleIIKeyboard F12/CTRL+BACKSPACE behavior, changed Apple II plus drive 2 to Alps mechanism, added volume control to Disk II, fixed Disk II not updating last track when closing OE, fixed OEEmulation not being destroyed when closing OE, fixed crash when changing device and device was selected, remade ctrl+option screen in Helvetica

+++ macosx/CanvasWindow.m
@@ -81,9 +81,9 @@
             
             windowRect = [self frame];
             NSRect contentRect = [[self contentView] frame];
-            float titlebarHeight = (float) (NSHeight(windowRect) -
-                                            ([self userSpaceScaleFactor] * 
-                                             NSHeight(contentRect)));
+            CGFloat titlebarHeight = (NSHeight(windowRect) -
+                                      ([self userSpaceScaleFactor] * 
+                                       NSHeight(contentRect)));
             
             NSRect screenRect = [[self screen] frame];
             screenRect.size.height += titlebarHeight;
